pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gravix"
version = "0.0.1rc0"
description = "Embedded, high-performance local-first engine for AI agent retrieval and memory"
readme = "README.md"
requires-python = ">=3.10,<3.15"
authors = [
    {name = "dichters", email = "dichterszergn@gmail.com"}
]
dependencies = [
    "ladybug==0.16.1",
    "openai==2.33.0",
    "loguru==0.7.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
gravix = "gravix.cli:main"
